Began Statistics Tracking via AudioTrackWithUser.
AudioTrackWithUser has dropped in for replacement anywhere AudioTrack was relevant for playing. AudioResultHandlers are no longer global and initiate for each query as to track the user and pass off to ProperTrackScheduler. Additionally, ProperTrackScheduler no longer holds just AudioTrack in the queue, but AudioTrackWithUser. ProperTrackScheduler also uses an AudioTrack to DiscordUser map as its events cannot be overridden to trigger AudioTrackWithUser on. Should not be a big deal as it is just storing a memory address, not a duplicate (unconfirmed). The TrackEventStatisticsRecorder has yet to be implemented anywhere, but is ready.